[FIX] helpdesk: prevent indexerror when creating helpdesk ticket
When the customer creates a helpdesk ticket and
the user of helpdesk team is deleted, a traceback will appear.

Steps to reproduce the error:
- Go to Settings > Users & Companies > Users > Create a user
- Go to Helpdesk > Configuration > Helpdesk Teams > Open Team A
  Select Assignment Method: Random > Select that user in the members > Save
- Delete that user
- Go to Helpdesk > All Tickets > Create a ticket > Select Team A

Error: A traceback appears:
"IndexError: list index out of range"

https://github.com/odoo/enterprise/blob/947b6ab9447aa069fd5edd150c932d9a0884a3cb/helpdesk/models/helpdesk.py#L502
When the related user is deleted, "member_ids" will be an empty list.
so when it tries to access the index of "member_ids",
it will lead to the above traceback.

sentry-5630888721

closes odoo/enterprise#70116

X-original-commit: d8da14d40eb7a7ec4187e705871e27dc1c9c9a9e
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>