[IMP] account_accountant: no fiscal year on child branches
Description of the issue this commit addresses:

We do not want the user to be able to create fiscal years on child companies.

Desired behavior after this commit is merged:

A new check in fiscal years raises an error when the user tries to create
a fiscal year from a child company.

no task-feedback

closes odoo/enterprise#67272

X-original-commit: 899b83405c27186c047a40041f04413c0cd2eec8
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>