[FIX] industry_fsm_sale: allow customer edition in portal subtasks su…
…bview

Steps:

Login as portal, and to to a task form, page sub-tasks.
Make Customer column visible.
Add a line and try to edit that field.

Issue:
You can't. But after saving, you can.

Cause:
The field is invisible if `allow_billable` and `is_fsm` are False.
A child task inherits both from its parent, but this one hasn't been
created yet.

Solution:
In this case, the condition is not needed. If we create the child and he
has both fields to False, it means its parent too. Yet the `partner_id`
column is invisible if it is the case.
So the solution is simple: remove the condition.

related: https://github.com/odoo/odoo/pull/174323
task-3713729

closes odoo/enterprise#67252

Related: odoo/odoo#174323
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>