[FIX] web_studio: apply datetime widget when unticking show time
After making changes in community PR: https://github.com/odoo/odoo/pull/161705

The datetime widget was not being applied when we dragged and dropped a datetime
field from the studio.

Steps to reproduce:

- Open the Sales module and toggle Studio.
- Drag and drop a datetime field into the list view.
- Try to tick and untick the Show time option.

Observed behavior:
When ticking and unticking the show time option, the time part is not hiding
before resetting the widget.

Expected behavior:
The time part should be hidden when we tick and untick the show time option
without resetting the widget.

After this commit, the datetime widget will be applied when we drag and drop a
datetime field from the studio and untick the show time boolean.

Task-3698841

closes odoo/enterprise#67988

X-original-commit: 92612b7ed3f64c670242cba54c13c2636136e730
Related: odoo/odoo#175743
Related: odoo/documentation#10603
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>