[FIX] account_accountant: tour can fail undeterministically
The tour test_tour_invoice_creation_from_reco_model fails sometimes.
The issue is that we try to click on the action to open the move
before the previous deletion has been in effect.
It seems a step_delay was introduced to mitigate the issue, but it
wasn't the right way.
We should instead ensure that the move is indeed deleted before
clicking on the next action.

runbot-72323

closes odoo/enterprise#67259

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>