[FIX] stock_barcode: digipad erasing does not work properly
Description of the issue/feature this PR addresses:

1. go to any operations of th barcorde app
2. go to a specific line
3. change the quantity thanks to the +1 / -1
4. try to erase using backspace or the erase button

Current behavior before PR:
We need several backspaces to erase the value

Desired behavior after PR is merged:
If the value is '3' for example, only 1 backspace should be needed.

Explanation:
Incrementing will set the precision to the value specified, which mean 3 is actually stored
as 3.00 on the component, but parsing it as a Number in _notifyChanges will remove these
trailing zeros.

One solution would be to pass the string to the ORM instead of passing it a number, but
this has the disadvantages to set the value as 3.00 visually (so the user still need to
delete multiple zeros.

The solution adopted removes the trailing zeros.

Side note: This does not fix the inconsitency where the value displayed is "3" when the
digipad value id "3.".

opw-4061351

closes odoo/enterprise#67660

X-original-commit: 3066ecd468fdf617a23c4c666edaed7546efacef
Signed-off-by: Steve Van Essche <svs@odoo.com>