[FIX] stock_barcode: GS1 packaging with lot
Steps to reproduce:
===================
- In Inventory settings, choose GS1 as barcode nomenclature and activate packages;
- Create a storable product tracked by lot;
- Create a lot for this product (e.g.: 12345);
- Create a second storable product tracked by lot;
- Create a new lot with the same name (12345) and a packaging on
  the product page with the barcode number 10000000240489;
- Go back to the barcode app and create a new delivery;
- San the barcode 01100000002404891012345
  => The line is added for the right product but the lot is missing.

What happens:
=============
When the GS1 barcode is parsed, we got a lot name and a packaging
barcode, thus, we will search a lot for the scanned lot name.
To avoid fetching the wrong lot in case multiple lots have the same name
(but different product), we give a filter when the lot is scanned
alongside a product.
The issue here is we didn't scan a product but a packaging and we forgot
to give the same filter in this case.

About this commit:
==================
This issue was indirectly fixed by following commit:
https://github.com/odoo/enterprise/commit/104399fb42f8b8a9c5a1432f8f61b54aedc2580d
That said, that commit doesn't test this flow, so adding the test/tour
is still relevant.

OPW-3894500

X-original-commit: 3e6c4ca75b028505b23ce4a3b4fc096207134bf2
Part-of: odoo/enterprise#68458
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>