[FIX] voip: speed up call to _format_call_activities
We notice that get_today_call_activities is a very frequent call that
spend a lot of time runing sql queries.

during 1h30 it was called 20515 times for an average of 719ms of sql times
on odoo.com database.

It was cause by _format_call_activities
which can be called with many hundreds of activities

before this commit for each activity
_mail_get_partners and mail_partner_format were called once per
activity. It take 581s for all the internal user of odoo.com database

after this commit
the method _mail_get_partners and mail_partner_format are called for an
entire batch as it's intended. The same computation for internal user
takes now 84s

closes odoo/enterprise#67718

X-original-commit: b31cf7dec220a125adbb860bc89165d5eba35ec3
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Thibault Francois <tfr@odoo.com>