[IMP] account_accountant: no fiscal year on child branches
Description of the issue this commit addresses:

We do not want the user to be able to create fiscal years on child companies.

Desired behavior after this commit is merged:

A new check in fiscal years raises an error when the user tries to create
a fiscal year from a child company.

no task-feedback

closes odoo/enterprise#67349

X-original-commit: 937136b89e6a737be2250bbc502d044742279685
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>