[IMP] frontdesk: add PWA support for frontdesks
With this commit, kiosks of the frontdesk app can be installed as
Progressive Web Apps. Using the install_kiosk_pwa action, a dialog
is shown when clicking on the button to open the kiosk, allowing to
see the link, or install the PWA from the dialog.

task-3953694

closes odoo/enterprise#66807

Related: odoo/odoo#173478
Related: odoo/upgrade#6280
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>