[IMP] spreadsheet: make chart clickable to see records
Task Description

This task adds the possibility to see the records related to a
bar/pie part in a Odoo chart in Spreadsheet (included via the
"Insert in spreadsheet" action)

We keep the original action xml id to be sure to redirect
to the correct view. (if you insert a chart of account.analytic.line,
we have to redirect to either the view in Accounting or in Timesheet
depending on where the chart comes from.

Related Task

closes odoo/enterprise#67353

Task: 3806411
Related: odoo/odoo#173470
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>