[FIX] whatsapp: prevent error when clicking on submit for approval in template
When attempting to create a new template in WhatsApp, leaving the 'Category'
field empty results in an error.

Steps to reproduce:
- Install the 'WhatsApp' module
- Create a 'WhatsApp Business Account'
- Create a new template and leave the 'Category' field empty
- Click on 'Submit for Approval'

Traceback: AttributeError
'bool' object has no attribute 'upper'

This error occurs at [1] because 'template_type' is not found in self.

This commit will fix the above error by raising an error if 'template_type'
not found in self.

[1]- https://github.com/odoo/enterprise/blob/d6e3dcc5a37c82a0c1bfc1267c4e088aedfe8632/whatsapp/models/whatsapp_template.py#L512

sentry-5642705166

closes odoo/enterprise#67861

X-original-commit: 08fc364893b306221b18a6a15ac31dc4055483ef
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>