[FIX] website_sale_stock_renting: add product name in payment error
__Current behavior before commit:__
When you check out your cart with a product not in the stock anymore you have an error message. There is no product name in this message so if you have multiple products in your cart you don't know which one to remove.

__Description of the fix:__
Add the product name in the error message, so it's easy to identify which product to remove of the cart.
The warning method is linked to sale.order.line so we can access to the product directly.

__Steps to reproduce the issue:__
- add products to your cart
- remove one or many products of the stock
- try to validate your payment

__linked to Odoo PR:__
https://github.com/odoo/odoo/pull/174615

opw-4016059 (upgrade issues)

closes odoo/enterprise#68546

X-original-commit: 63e2a3ec4a3565c16a0322d5bdafc3cf7ee6d7d6
Related: odoo/odoo#177043
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>