[FIX] account_budget: merge queries and multiple fixes
Currently, there are 2 separate queries that can be merged. On the one
hand, the committed and achieved amounts on the `budget.line` are
obtained thanks to a compute that executes a big SQL query. On the other
hand, there exists a `_table_query` on the non-stored `budget.report`
model that allows to obtain the lines related to a `budget.line`. The
sum of these lines gives us the committed and achieved amounts.
Consequently, we can use the same query to get these amounts and lines.

To do so, we need to add a Many2one on the `budget.report` which allows
to filter and groupby the `budget.line`.

In addition, this commit fixes multiple issues in the remaining
subqueries. Remove a useless over-complex lateral join. Fix the "join"
condition between `budget.line` and `account.analytic.line`: retrieve
only the `account.analytic.line` having a matching
`account.analytic.account` set on one of their plans. The added
parenthesis are important to ensure the correct precedence.

Finally, we need to flush the models used by the SQL query to ensure the
database is up-to-date.

X-original-commit: 6116063f380996c2970ba6ab43aac492e1b04e12
Part-of: odoo/enterprise#71430
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>