[IMP] sale_subscription_external_tax: current date for subscriptions
By default, we send date_order to Avatax when asking to calculate
taxes. Avatax calculates the right taxes for that point in time. For
non-recurring orders this is ok, but subscriptions are paid multiple
times and date_order will remain the confirmation date of the
subscription.

This leads to confusion as taxes change over time (e.g. tax laws
change or an exemption certificate is added with an effective date
after date_order). Because customers are charged based on invoices we
always charge them correctly, but the subscription total may not match
what they're being charged.

To solve this, always send the current date for subscriptions. This
way customers can accurately see what they'll be charged in the
portal.

task-4078473

closes odoo/enterprise#67817

X-original-commit: 3bfbdf23df9cf4e8ebb3fa10137c7eda30597bd8
Signed-off-by: William André (wan) <wan@odoo.com>