[FIX] appointment: fix traceback when user gives existing email in guest email
This traceback arises when the user adds an existing attendee email
in guest email, while adding a guest.

To reproduce this traceback:

1) Install website_appointment
2) Choose any appointment from the website
3) Click on 'Appointment Type' in the right corner
   and on the options page click 'Allow Guests'
4) Get back to the website and select one user, date & time
5) Now Add Guests and give Email value as email in Guests
6) Confirm Appointment

Error:- "ValueError: Expected singleton: calendar.attendee(4762, 4763)"

When the user adds a guest with the current user email, multiple attendees
will be created with the same partner ID in a calendar event (see below)

https://github.com/odoo/enterprise/blob/1aff50d91fad45221377591b80ca8b1df2ecb593/appointment/models/appointment_type.py#L944-L955

This leads to the above traceback from here

https://github.com/odoo/enterprise/blob/1aff50d91fad45221377591b80ca8b1df2ecb593/appointment/controllers/calendar.py#L103

This commit will resolve the issue by filtering out the existing attendees while
creating a new attendee from the guest email.

sentry-5646921453

closes odoo/enterprise#68190

X-original-commit: 9a502b8dfcded175cd8f089f0530682738681b8c
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>