[IMP] sale_subscription: remove unused argument
In this commit, remove the 'message_class' argument from the subscription portal
as it is no longer used.

when set default payment token from portal for subscription the value passed in
url is not used anymore. If not using that value is intentional then this
changes need to be applied. otherwise we should not remove message_class
argument to display this message on portal.

task-4077522

closes odoo/enterprise#67530

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>