[FIX] iot: ensure at least one device is selected
Currently, when printing reports using the iot, if we don't select a device
when the wizard prompts ut to do so, nothing will be printed and it will be
saved in the browser storage that no printer should print this report.

Steps to reproduce:
-------------------
* Connect the IOT to the db
* In developper mode go to: **Settings** > **Technical** > **Actions** > **Reports**
* Select `PDF Quote` for example and associate a printer
* Go to the **Sale** App and select any quotation
* Select the gear icon
* Select **Print** > `PDF Quote`
* Select `Print` without selecting a printer
> Observation: Nothing is printed
* Select **Print** > `PDF Quote`
> Observation: Nothing is printed and the prompt does not appear

Why the fix:
------------
The idea behind this fix is to make it mandatory to choose at least one of
the devices on the wizard prompt.

If at least one printer is selected we continue with the printing. If not,
we put a notification on screen and don't do anything else.

opw-4019575

closes odoo/enterprise#67579

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>