[IMP] stock_barcode: sound traceback
With some OS/web browser, it is not allowed to automatically play sounds
with no user interaction (typically: a button on which the user has to
click to play the sound) which means the `play` promise could be
rejected. When this happens, an Odoo traceback is thrown because of the
uncaughted rejected promise.
Add a `catch` for the promise to avoid that.

task-3929927

Part-of: odoo/enterprise#63727
Related: odoo/odoo#167942
Related: odoo/upgrade#6249
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>