[FIX] website_sale_renting: disable date picker if exisiting rental period
Issue
-----
[website_sale_renting]
In case there is already a rental product in the cart, it is still possbile
to open the date range picker by clicking the calendar button and selecting
a new period, which will have no effect as the rental period can only be
changed from the cart.

Change
-----
Prevent date picker opening but keep the calendar button as it contains
a help message.

opw-4076088

closes odoo/enterprise#67645

Signed-off-by: Valentin Chevalier <vcr@odoo.com>