[FIX] planning: fix traceback when user gives invalid hour value to the end date
This tracaback arises when the user gives invalid hour value to the end date
while creating or updating  a planning template.

To reproduce this issue:

1) Install Planning
2) Create a new shift template from planning configuration
3) Give an invalid hor value to the end date (greater than 24)
4) Try to save the record.

Error:- "ValueError: hour must be in 0..23"

When the user gives an invalid hour value to the end date and tries to save it,
a compute method triggers in which hours are used to calculate the end time.

https://github.com/odoo/enterprise/blob/14b37f4141f1c73712d2c5ce16e5fc815e2a6a9b/planning/models/planning_template.py#L37-L38

Which leads to the traceback.

Already a condition is there to check the invalid hours for start time.
By just modifying the existing condition we can resolve this issue.

sentry-5655372261

closes odoo/enterprise#68304

X-original-commit: e9dafe80f25e44c0a9a771eebfce961e7d4e5956
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>