[IMP] pos_restaurant_appointment: imp booking ux
In this commit we make the "manage bookings" gantt view appear as an
entire screen, instead of a dialog.

This makes the gantt view more clear and also changes the undesirable
behavior of the form view for editing events that would close the gantt
view behind it on save.

closes odoo/enterprise#67855

Task: 4089477
X-original-commit: da6f6786e5ec665bd6bd96b61efd80a2dd7978fa
Related: odoo/odoo#175489
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>