[FIX] l10n_latam_invoice_document: prevent blocking foreign document number
Steps to reproduce:
[l10n_ec]
- Create a credit note from the Bill journal
- Set a foreign customer
- Set a customized document number

Issue:
An error will be raised saying that the format is not correct

But, as defined by VBE, "If a Credit Note is created from a Vendor Bill and the partner_id != "EC", [we should] allow the user to allocate any number without following the EC format."

Solution:
When we call `_format_document_number` we don't have any information about the initial move. Instead of using a context or adding new fields, we add a hook in which we can specify
certain conditions to bypass the document check/formatting for localisations.

We add to add `partner_id` to avoid the flow in wich the user set a foreign customer and a custom document type to change back the customer to an ecuadorian one; hence the second part of the test

opw-3993305

closes odoo/enterprise#68775

X-original-commit: 39a55d9d915224fb764c4ff966b5fcef9bf57d9c
Related: odoo/odoo#177597
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>