[FIX] l10n_latam_invoice_document: prevent blocking foreign document number
Steps to reproduce:
[l10n_ec]
- Create a credit note from the Bill journal
- Set a foreign customer
- Set a customized document number

Issue:
An error will be raised saying that the format is not correct

But, as defined by VBE, "If a Credit Note is created from a Vendor Bill and the partner_id != "EC", [we should] allow the user to allocate any number without following the EC format."

Solution:
When we call `_format_document_number` we don't have any information about the initial move. Instead of using a context or adding new fields, we add a hook in which we can specify
certain conditions to bypass the document check/formatting for localisations.

We add to add `partner_id` to avoid the flow in wich the user set a foreign customer and a custom document type to change back the customer to an ecuadorian one; hence the second part of the test

note fw-port 18.0:
As from 18.0 there is a warning `condition_to_sql: ignored ('l10n_ec_withhold_invoice_id', '=', NewId), did you mean ('l10n_ec_withhold_invoice_id', 'in', recs.ids)`
The `onchange` dependency change triggers the compute. `Self` being a `NewId`, the search's domain `('l10n_ec_withhold_invoice_id', '=', NewId.id) raising the warning and canceling the sql statement (leading to potential error; no flow found)
https://github.com/odoo/odoo/blob/686628f55f6eca35d8caa500157a1f0b7353c4e4/odoo/models.py#L3183-L3185

Using `('l10n_ec_withhold_invoice_id', 'in', invoice.ids) would do the trick as the property `ids` uses `origin_ids` retrieving the correct id (if any) even for a NewId object
https://github.com/odoo/odoo/blob/686628f55f6eca35d8caa500157a1f0b7353c4e4/odoo/models.py#L296-L300

BUT as we are in a compute and the initial code was searching in a loop, I also took out from the loop `search`

opw-3993305

closes odoo/enterprise#70673

X-original-commit: 00420d2e228109adbed5c6e410208b4572fa38e1
Related: odoo/odoo#181450
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>