[FIX] account_sepa{,_pain_001_001_09}: add user error for missing UETR
When switching a bank journal to the new SEPA Pain version, existing
payments in the batch may lack required data, such as the UETR field.

Steps to Reproduce
------------------
1. Install `l10n_be`, `account_accountant`, and
   `account_sepa_pain_001_001_09`.
2. In the settings, enable 'SEPA Credit Transfer (SCT)'.
3. Switch to a Belgian company.
4. Set up an IBAN bank account on the bank journal.
5. Ensure a partner has an IBAN bank account defined.
6. Create and confirm an outgoing payment for that partner, setting the
   payment method to 'SEPA Credit Transfer'.
7. In the Outgoing Payments tab of the bank journal, set the SEPA Pain
   Version to 'New generic version'.
8. Create a batch payment with the payment you just created.
9. Attempt to validate it.

Expected Result
---------------
The payment should be validated successfully.

Actual Result
-------------
A traceback occurs: `TypeError: Argument must be bytes or unicode, got
'bool'`

Cause
-----
The payment in the batch was created before the SEPA PAIN version was
updated on the bank journal. Consequently, the payment is missing some
necessary data (specifically, the UETR field is not set) required for
the new SEPA version.

opw-3973374

closes odoo/enterprise#67451

X-original-commit: 607273f5c7790ced1703d6cc90bbff58b772ff53
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>