[FIX] account_auto_transfer: round automatic transfer lines
Steps to reproduce:
- Install Accounting
- Go to "Accounting / Configuration / Accounting / Chart of Accounts"
- Create an account (e.g. Test Account)
- Go to "Accounting / Accounting / Miscellaneous / Journal Entries"
- Create a journal entry:
  * Journal Items:

        Account    |    Debit    |    Credit
    --------------------------------------------
     Test Account  |    $0.00    |    $410.34
         [any]     |   $410.34   |     $0.00

- Post the journal entry
- Go to "Accounting / Accounting / Management / Automatic Transfers"
- Create an automatic transfer:
  * Origin Accounts: Test Account
  * Automated Transfer:

        Percent (%)    |    Destination Account
    -----------------------------------------------
           15.00       |           [any]
           42.50       | [any, but a different one]
           42.50       | [any, but a different one]

- Activate the automatic transfer
- Compute transfer

Issue:
A UserError is raised while trying to create a journal entry because
of a $0.01 difference between the total credit and the total debit.

Cause:
When the amounts of each line are computed from the percentage, they are
not rounded.
These amounts are used to create the journal entries generated by the
automatic transfer.
However, each created journal entry line is rounded, which can generate
a rounding difference.

Solution:
Directly round the amounts when they are computed from the percentage to
be sure that the amounts used for the computation and the ones that will
be set in the journal entries are the same.

opw-3998808

closes odoo/enterprise#67946

X-original-commit: 98e72eed05dc0a5eb8080debe88cba96bb0a7a14
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>