[FIX] appointment: count booking capacity on manual treshold
A coding mistake was added in d53aae967c5b44b2ca1ed716e90e065fbbe938d4
Indeed, for resource bookings, we decrease asked capacity
to allocate the full reserved capacity to available resources.
However, it is decreased directly.

This means that event values have a 0 asked_capacity, leading
in turns to an inexact use of _get_default_appointment_attendee_status,
used to determine whether the status should be 'accepted' or
'needsAction' depending on the manual treshold percentage if set,
asked capacity and existing bookings.

Indeed, the asked_capacity will not be counted when computing the
total reserved percentage. In practice, this means that events
crossing the treshold will be automatically 'accepted' instead of
'needsAction'

The fix creates another variable to avoid the issue and adds a test.

Task-4084727

closes odoo/enterprise#67626

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>