[FIX] sale_renting: correctly copy rental order lines
Problem:

When the user duplicates a rental order, the rental order lines were not
considered rental lines anymore after duplication.
The field was not configured to be copied on duplication, and was
recomputed on the new lines, but the `in_rental_app` context was not
given to the copy call, so all lines were considered standard lines.

Solution:

Set `is_rental` copy=True and readonly=False
to ensure the value is properly copied on duplication.

Steps to Reproduce on Runbot:
1. Install Rental app
2. View the rental orders in list view
3. Select a rental order with a status and duplicate
4. Observe that duped rental order has no rental_status
when you confirm it

opw-4055493

closes odoo/enterprise#68397

X-original-commit: 53f73e3ba78b77a4937b2419809b5f810b20ded7
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>