[IMP] account_reports: journal groups
This commit will change some things in the filter for journal groups, the
display when changing journal group has been modified, by default when coming on
a report when having a journal group it's automatically applied and when
checking the journal items from a report with journal group selected, the domain
is changed.

Also fixed a but for the domain where the journal_types gave the journal types
of the excluded journals and not the remaining one from the journal_group.

Task: 4058751
Part-of: odoo/enterprise#67611
Related: odoo/odoo#174988
Related: odoo/upgrade#6350
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>