[FIX] documents_account: fix documents_account_tour tour
From https://github.com/odoo/odoo/pull/174548, tooltip has been renamed
in tooltipPosition. This conversion for this tour has been missed.
In https://github.com/odoo/odoo/pull/171756, the step schema is
validated when a tour is added to the registry.
This is why this bug was detected.

task~3974087

closes odoo/enterprise#67582

Related: odoo/odoo#171756
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>