[FIX] web_studio: prevent adding field in view when there is no node
Currently, an error is generated when the user tries to add a new
'One2Many' field in the model.

Steps to produce:
1. Create a new app and new model with studio
2. Try to add the 'One2Many' field in view >>> error occurs

error: `KeyError: 'node'`

This is because when we try to add 'One2Many' in view, there is no 'one2many'
field available in the model, so it tries to add a new field without a node.

This commit will fix the above issue by preventing adding structure
when there is no node (_operation).

sentry-5656743440

closes odoo/enterprise#68179

X-original-commit: d59cbb3cdada5faadfb3982cc26d73797ed4f413
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: ANSARI MAHAMADASIF (maan) <maan@odoo.com>