[FIX] industry_fsm_sale: apply fiscal position on employee rate
Steps to reproduce:

- Create a fiscal position that converts 19% tax-inclusive into 19%
tax-exclusive.
- Create a product with 19% tax inclusive.
- Create a contact and assign the above fiscal position.
- Navigate to a field service project and add an employee with a
tax-included rate in the employee mapping.
- Create a new field service task for the customer with the assigned
fiscal position and add one hour for the employee in the mapping.
- Add the product with 19% tax included to the task.
- Mark the task as done.

Current Behavior:
- The product unit price is calculated excluding tax.
- The service unit price remains calculated as tax-included.

Expected Behavior:

Both product and service unit prices should be calculated consistently
as tax-excluded when the tax is set to 19% exclusive.

Cause of the issue:

When creating a new sale order line if the unit price is not specified
"product_id_change" is called which set the correct unit price after
applying the fpos by calling `_update_taxes` on the line:
https://github.com/odoo/odoo/blob/a09e0c651ea54faa3b41e6012b6d62f143519f07/addons/sale/models/sale_order_line.py#L617-L628
But in our case we are setting the `unit_price` to the employee rate.
here:
https://github.com/odoo/enterprise/blob/e23e01337faeec652d400618cf14aaf3d6d47ef2/industry_fsm_sale/models/project_task.py#L408-L412
and the taxes are not taken into account.

Fix:

The unit price of the employee rate should be updated to reflect the
changed taxes from fpos. However, since `_update_taxes` relies on the
unit price of the product rather than on the one given to the line:
https://github.com/odoo/odoo/blob/a09e0c651ea54faa3b41e6012b6d62f143519f07/addons/sale/models/sale_order_line.py#L685
we can not call the `_update_taxes` directly to update the sol.

opw-3853196

closes odoo/enterprise#68635

X-original-commit: be757e4d4d1a3b3b913fd48b423bf259dfcf2b05
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Co-authored-by: waha <waha@odoo.com>