[FIX] l10n_ec_edi: display the unit price without taxes on ecuadorian invoice
The Ecuadorian law requires the unit price on the XML invoice and on the PDF to be the same, this is currently not the case with taxes included in the price.

Steps to reproduce:
- Install the 'l10n_ec' module and switch to an Ecuadorian company
- Change the 'IVA 15% (411, Servicios)' Tax to be included
- In Accounting create a new Customer Invoice
- Select 'Instituto Ecuatoriano' as Customer
- Add a line with the included tax set before
- Select 'Sin utilization del sistema financiero' as Payment Method
- Confirm
- On the blue popup at the top of the page, click 'process now' to get the XML in the chatter
- Click preview to get the PDF
- The unit prices are not the same

Cause:
Usually on the PDF the unit price with included taxes is displayed.

Solution:
Create a xpath to compute the unit price in the same way as the XML.

opw-4078177

closes odoo/enterprise#68363

X-original-commit: b07eca431842081d3abc1040e408d475f9ee8d4f
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>