[FIX] website_helpdesk: fix `s_call_to_action` design.
Commit[1], which is part of the Odoo 18 snippets redesign aims to align
all the occurrences of the `s_call_to_action` snippet across
the front-end.

It appears that this commit did not take into account the fact that
a second button can be xpath next to the first one.

To ensure this scenario does not cause any issue, we restore the flex
layout already existing before.

Commit[1]: 53c0592f367572db5ceff49db2f4363b7306a102

task-4086686

closes odoo/enterprise#67728

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>