[FIX] l10n_mx_edi: Override peppol fields computation
Version: 17.0+

Current Behavior:
Peppol fields are automatically populated based on the `vat` field of the res.partner record.
With l10n_mx localization installed the `vat` field is based on the first two letters of the
first name and the first letter of the last name. However, the peppol fields with be automatically computed if the first two
letters are similar to country codes.

Purpose of this PR:
This PR overrides the peppol computation so that the peppol fields are emptied when l10n_mx_edi is installed.
These fields are unused in l10n_mx localization so they can be omitted.

Steps to Reproduce on Runbot:
1) Install accounting and contacts
2) Create a contact with l10n_mx vat format (i.e. ESPN0101011E2)
3) Navigate to the Accounting Tab of the partner record and find that the peppol fields are filled.

Notes:
In the current behavior, the user will have to manually empty the peppol fields to avoid receiving the error of incorrect peppol format

opw-4074094

closes odoo/enterprise#68257

X-original-commit: 3f1c73f9e24c08afe8b841b2ff13d4610cdc49ed
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>