[IMP] account_reports: general ledger communication
In this commit https://github.com/odoo/odoo/pull/175170/commits/0068c71a864407ac6ee621d0c6c3519ec05fc30e
we changed the way the name of the move line is computed for receivable.
In the general ledger we use the name of the move line for the communication
columns, so it needed to be adapted.

Task: 4058751
Part-of: odoo/enterprise#67611
Related: odoo/odoo#174988
Related: odoo/upgrade#6350
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>