[FIX] sign: template label overlap and search panel hiding
Before this commit, the bottom right label in sign templates which included
the hourglass icon was overlapping the action buttons and the search panel
in the left position of the screen was occupying a big part of the screen when
the width was small, hiding the sign templates and sign documents in the bottom of the page.

After this commit, the bottom right label does not overlaps the action buttons
anymore and when the screen is small (in width) we hide the search panel
for showing the sign templates and sign documents correctly.

task-4077791

closes odoo/enterprise#67700

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>