[FIX] hr_payroll_expense: Reset expense sheet to approved if slip is reset to draft
If you generate a payslip with some expenses into it and set the
payslip as "Done", the state of the Expense will be set to 'Done' too.
If you set the payslip back to draft, it won't be possible to create draft
entry again cause the linked expense state didn't reset back to 'approved'.
Instead, it remained as 'Done'.

To fix the issue, when the payslip is cancelled (before it is set back
to draft), set the state of all linked expenses back to 'approved' and
reverse all the account moves that were created when the expenses state
was set to 'Done'.

task-4045573

closes odoo/enterprise#67779

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>