[FIX] industry_fsm_report: fix industry_fsm_tour to run independently
Currently, the tour is written according to the demo data.

The selector .o_form_sheet div[name] input will match if the work template is
'Device Installation and Maintenance' .
This template is available in the demo. If the demo loads, the default
'Device Installation and Maintenance' will be set via the post hook. When the
demo does not load, we will have the default template instead.
That's why we used the `.o_form_sheet p.oe-command-temporary-hint` class in the
OR condition.

task-4086026

Part-of: odoo/enterprise#67846
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>