[FIX] appointment: fix traceback when computing leave resource
How to reproduce
=================
Step 1) Create an event on resources and set the start time = end time
Step 2) Click on the edit button
-> Traceback
video ref - https://youtu.be/csFkJjO6hMQ

Technical
=========
When we set the start = stop then we won't be able to find the interval so we
failed to assign the value of the field due to this traceback occurs.

After this commit
==================
Now the traceback will not occur.

Task-4010880

closes odoo/enterprise#69221

X-original-commit: bf0cac36a8153620dccfbc3cad41f864f7521837
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Shaan Thakkar (thsh) <thsh@odoo.com>