[IMP] account_asset: replace one2many to many2many
In this commit, we will change the asset_model field to a many2many to be able
to generate multiple assets from one invoice.
We also added a stat button on the asset to see the linked asset and check if
one of the linked asset has been confirmed.
Added the filter journal

Task: 4058751
Part-of: odoo/enterprise#67611
Related: odoo/odoo#174988
Related: odoo/upgrade#6350
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>