[FIX] marketing_automation: fix kanban open action
Purpose
=======
Fix the "open record" action for the activities kanban in the marketing campaign form.
The record can be opened when clicking anywhere on the kanban article,
it should only be opened when clicking on the title of the acitvity.

Specification
=============
Following this commit odoo/odoo@d52fc5ae69a9ca55df22f416a26c1ea5f5484ac9 the kanban arch parser
has been updated changing the condition to open the record from this 'this.allowGlobalClick'
to this 'this.props.archInfo.canOpenRecords'.

It was previously only opening the record when the user clicked on the element
with the 'oe_kanban_action' class (i.e. the activity title) but is now opening
the record no matter where the click is on the kanban article.

Fixing the issue by adding the new 'can_open' attribute on the kanban and setting
it to false bringing back the old behavior.

Also adding the 'user-select-none' class on the 'Add child activity' button so
that the cursor isn't changed to the selector one when hovering the text.

Task-4089001

closes odoo/enterprise#68904

X-original-commit: c4c3e630d990f96e3c794dc24ff49c1f53c859ae
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>